chore: move helpers to testing/util.go, delete testing/tenant.go
